Issue 83: Moved Hadoop wordcount example into a specific package.
