+++ hadoop-connector-examples/src/main/java/io/pravega/example/hadoop/wordcount/TextSerializer.java
@@ -8,7 +8,7 @@
  *     http://www.apache.org/licenses/LICENSE-2.0
  */
 
-package io.pravega.example.hadoop;
+package io.pravega.example.hadoop.wordcount;
 
 import io.pravega.client.stream.Serializer;
 import lombok.EqualsAndHashCode;
